+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [TBD]
-* Awaiting MFA Delegate now automatically returns the AuthMethods to be used when calling MFA Challenge
+* Awaiting MFA Delegate now automatically returns the AuthMethods to be used when calling MFA Challenge #2687
+* Prevent double conversion of MSALInternal error #2688
 
 ## [2.3.0]
 * Fix automation tests (#2673)