+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [2.4.0]
+* Prevent double conversion of MSALInternal error #2688
+
 ## [2.3.0]
 * Fix automation tests (#2673)